Add aarch64 version of unpack_bitmap
Only uses 4 bits for lookup table because only 128 bit registers are
supported. In theory, aarch64 supports instructions that write two
registers at once, but Rust doesn't seem to have the instrinsics for
those right now.

The upside is that the lookup table is really small (32 Bytes).

The downside is that it's over 2x slower than the AVX2 version (65ns vs
165ns). But it's still faster than the scalar version except for very
sparse bitmaps:

    density     scalar      simd      ratio
      0.0625    132ns       164ns     0.8x (-1.25x)
      0.125     264ns       164ns     1.6x
      0.25      522ns       164ns     3.2x
      0.5      1053ns       164ns     6.4x
      0.75     1588ns       165ns     9.7x
      0.9      1910ns       167ns    11.6x